add todo and move test
