+++ google-cloud-storage/src/main/java/com/google/cloud/storage/GrpcStorageOptions.java
@@ -601,6 +601,7 @@ public final class GrpcStorageOptions extends StorageOptions
     @Override
     public GrpcStorageOptions build() {
       GrpcStorageOptions options = new GrpcStorageOptions(this, defaults());
+      // TODO: Remove when https://github.com/googleapis/sdk-platform-java/issues/2911 is resolved
       if (options.getUniverseDomain() != null) {
         this.setHost("https://storage." + options.getUniverseDomain());
         return new GrpcStorageOptions(this, defaults());
