+++ google-cloud-storage/src/test/java/com/google/cloud/storage/StorageOptionsBuilderTest.java
@@ -69,17 +69,6 @@ public final class StorageOptionsBuilderTest {
         () -> assertThat(rebuilt.hashCode()).isEqualTo(base.hashCode()));
   }
 
-  @Test
-  public void testGrpcUniverseDomainMatchesHost() throws Exception {
-    Storage storage =
-        StorageOptions.grpc().setUniverseDomain("my-universe-domain.com").build().getService();
-    assertAll(
-        () -> assertThat(storage.getOptions().getUniverseDomain().equals("my-universe-domain.com")),
-        () ->
-            assertThat(
-                storage.getOptions().getHost().equals("https://storage.my-universe-domain.com")));
-  }
-
   private static class MyStorageRetryStrategy implements StorageRetryStrategy {
 
     @Override
